Restart service on changes to binary or systemd unit file
If there are changes to the binary it usually means that a new version
of sonarr has been installed, so the service should be restarted to use
this new version.

Similarly if there are changes to the systemd unit file it implies that
how sonarr is started has changed, and the service needs to be restarted
to use the changes.