Update the MediaInfo library package name
The libmediainfo0 package has been renamed to libmediainfo0v5 as of
Debian 9/stretch. We could maintain backwards compatibility with
Debian 8/jessie, but I am only using this formula with the latest stable
version of Debian.

If someone wanted to implement support for multiple Debian releases I
would be happy to merge the changes.